Версия MVC 0.5.6 Незначительные исправления и дополнения
